+++ CHANGELOG.md
@@ -503,3 +503,5 @@
 * Hit puffs and entity controllers now use custom property drawer for hit material
 * RpgDefaultValues is now a non-static class and implementes IRpgDefaultValues interface
 * Difficulty multipliers for player movement, attack, etc is now handled in RpgValues calculations
+* RpgDefaultValues can be overridden in its entirety by providing an IRpgDefaultValues implementation with RpgDefaultValuesOverrideAttribute (this is used for legacy back compat where we need the old calculations)
+*